another cleanup (#19)
* remove redundant morton_scalar

* remove use of _morton3_scalar

* cleanup prints

* replace Float64(0.0) by simply 0.0

* rename is_mt to parallel

* remove extra ray_tri intersect function. use common version.

* cl

* move intersect code to one file

* fix type

* shorten _write_child_bounds

* update readme

+++ bajo/core/bvh/__init__.mojo
@@ -71,7 +71,7 @@ def trace_bvh_shadow(bvh: BinaryBvh, rays: List[Ray]) -> Int:
 
 
 def trace_bvh_primary(bvh: BinaryBvh, rays: List[Ray]) -> Float64:
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = 0
     for i in range(len(rays)):
         var ray = rays[i].copy()
